fixing typecasting problems

+++ t5/ClothingShop/SummerDiscountStrategy.java
@@ -6,10 +6,10 @@ public class SummerDiscountStrategy implements DiscountStrategy
         long basePrice = clothing.getBasePrice();
         Season season = clothing.getSeason();
         //SPRING, SUMMER, FALL, WINTER
-        if(season == Season.SPRING) return 0.6*basePrice;
-        if(season == Season.SUMMER) return 0.5*basePrice;
-        if(season == Season.FALL) return 1*basePrice;
-        if(season == Season.WINTER) return 0.7*basePrice;
+        if(season == Season.SPRING) return (long) 0.6*basePrice;
+        if(season == Season.SUMMER) return (long) 0.5*basePrice;
+        if(season == Season.FALL) return (long) 1*basePrice;
+        if(season == Season.WINTER) return (long) 0.7*basePrice;
 
     }
 }
